requirements.txt
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.5
pandas==0.23.3
Pillow==5.2.0
progressbar2==3.38.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
reportlab==3.5.0
six==1.11.0
